requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.4.26
    # via sentry-sdk
django==5.2.2
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-allauth
    #   django-classy-tags
    #   django-cors-headers
    #   django-filter
    #   django-recaptcha
    #   django-sekizai
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   easy-thumbnails
django-admin-sortable2==2.2.8
    # via -r requirements.in
django-allauth==65.9.0
    # via -r requirements.in
django-classy-tags==4.1.0
    # via django-sekizai
django-cors-headers==4.7.0
    # via -r requirements.in
django-dbsettings==1.3.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-recaptcha==4.1.0
    # via -r requirements.in
django-sekizai==4.1.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
drf-spectacular-sidecar==2025.6.1
    # via -r requirements.in
easy-thumbnails==2.10
    # via -r requirements.in
freezegun==1.5.2
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.24.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown==3.8
    # via -r requirements.in
pillow==11.2.1
    # via
    #   -r requirements.in
    #   easy-thumbnails
psycopg2==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via freezegun
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.29.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.14.0
    # via referencing
tzdata==2025.2
    # via django
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.4.0
    # via sentry-sdk
